fix: tests and mocks
