--- babel.config.js
@@ -1,15 +1,0 @@
-module.exports = api => {
-  api.env('production');
-
-  return {
-    presets: [
-      '@babel/env',
-      '@babel/preset-typescript'
-    ],
-    env: {
-      production: {
-        presets: ['minify']
-      }
-    }
-  };
-};
