+++ package.json
@@ -2,14 +2,15 @@
   "name": "@ketch-sdk/ketch-tag",
   "version": "0.0.1",
   "description": "Ketch TypeScript/JavaScript Web Tag",
-  "main": "dist/ketch-tag.js",
-  "module": "dist/ketch-tag.esm.js",
-  "jsnext:main": "dist/ketch-tag.esm.js",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
   "scripts": {
-    "build": "rollup -c",
-    "test": "jest --runInBand",
-    "format": "npx prettier --write \"**/*.{ts,tsx,yml,yaml}\"",
-    "docs": "npx typedoc --excludeInternal src/pure.ts"
+    "build": "ncc build --minify --license licenses.txt src/index.ts",
+    "lint": "eslint src/**/*.ts",
+    "test": "jest --runInBand --passWithNoTests",
+    "format": "prettier --write \"**/*.{ts,tsx,yml,yaml}\"",
+    "format-check": "prettier --check '**/*.ts'",
+    "docs": "typedoc --githubPages true --excludeInternal src/pure.ts"
   },
   "repository": {
     "type": "git",
@@ -17,42 +18,34 @@
   },
   "author": "Ketch (https://www.ketch.com/)",
   "license": "MIT",
-  "homepage": "https://github.com/ketch-sdk/ketch-tag#readme",
+  "homepage": "https://github.com/ketch-sdk/ketch-tag",
   "bugs": {
     "url": "https://github.com/ketch-sdk/ketch-tag/issues"
   },
   "dependencies": {
-    "@ketch-sdk/ketch-plugin": "^0.3.3",
-    "@ketch-sdk/ketch-web-api": "^0.2.0",
-    "uuid": "^8.3.2"
+    "@ketch-sdk/ketch-web-api": "^1.0.2",
+    "loglevel": "^1.8.0",
+    "uuid": "^9.0.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.18.13",
-    "@babel/eslint-parser": "^7.18.9",
-    "@babel/preset-env": "^7.18.10",
-    "@babel/preset-typescript": "^7.18.6",
-    "@rollup/plugin-babel": "^5.3.1",
-    "@rollup/plugin-node-resolve": "^13.3.0",
-    "@rollup/plugin-replace": "^4.0.0",
-    "@types/jest": "^28.1.7",
+    "@ketch-sdk/ketch-types": "^0.2.4",
+    "@types/fetch-mock": "^7.3.5",
+    "@types/jest": "^29.2.0",
     "@types/uuid": "^8.3.4",
-    "@typescript-eslint/eslint-plugin": "^5.34.0",
-    "@typescript-eslint/parser": "^5.34.0",
-    "babel-jest": "^28.1.3",
-    "eslint": "^8.22.0",
+    "@typescript-eslint/eslint-plugin": "^5.40.1",
+    "@typescript-eslint/parser": "^5.40.1",
+    "eslint": "^8.26.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-jest": "^26.8.7",
+    "eslint-plugin-jest": "^27.1.3",
     "eslint-plugin-prettier": "^4.2.1",
-    "jest": "^28.1.3",
-    "jest-environment-jsdom": "^28.1.3",
-    "jest-junit": "^14.0.0",
-    "loglevel": "^1.8.0",
+    "jest": "^29.2.1",
+    "jest-environment-jsdom": "^29.2.1",
+    "jest-fetch-mock": "^3.0.3",
+    "jest-junit": "^14.0.1",
     "prettier": "^2.7.1",
-    "rollup": "^2.78.1",
-    "rollup-plugin-typescript2": "^0.33.0",
-    "ts-jest": "^28.0.8",
-    "typedoc": "^0.23.15",
-    "typescript": "^4.7.4"
+    "ts-jest": "^29.0.3",
+    "typedoc": "^0.23.18",
+    "typescript": "^4.8.4"
   }
 }
